+++ app/(pages)/(content)/sponsors/create/page.tsx
@@ -1,10 +1,11 @@
 'use client';
 
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import Link from "next/link";
 import { useAuth } from "@/contexts/auth-context";
-import { createSponsor } from "@/utils/sponsorsCreate.http";
-import { postMedia } from "@/utils/media.http"
+import { createSponsor } from "@/utils/sponsor.http";
+import { postMedia } from "@/utils/media.http";
+
 export default function SponsorFormPage() {
     const [companyName, setCompanyName] = useState('');
     const [responsibleName, setResponsibleName] = useState('');
@@ -14,32 +15,42 @@ export default function SponsorFormPage() {
     const [logoFile, setLogoFile] = useState<File | null>(null);
     const [bannerFile, setBannerFile] = useState<File | null>(null);
     const [reason, setReason] = useState('');
-    const { authToken, user } = useAuth();
-    const [logoPreview, setLogoPreview] = useState<string | null>(null);
-    const [bannerPreview, setBannerPreview] = useState<string | null>(null);
-    const [logoId, setLogoId] = useState<number | null>(null);
-    console.log(user?.role);
+    const { authToken } = useAuth();
 
     const handleSubmit = async (e: React.FormEvent) => {
         e.preventDefault();
-        const formData = new FormData();
-        formData.append('email', email);
-        formData.append('wantsLogo', wantsLogo.toString());
-        formData.append('wantsBanner', wantsBanner.toString());
-        formData.append('reason', reason);
+
         if (!authToken) {
-            throw new Error("El token de autenticación es requerido");
+            alert("El token de autenticación es requerido");
+            return;
         }
+
         try {
+            let logoId: number | undefined;
+            let bannerId: number | undefined;
+
+            if (wantsLogo) {
+                if (!logoFile) {
+                    alert("Debe seleccionar un archivo para el logo.");
+                    return;
+                }
+                logoId = await postMedia(logoFile, authToken);
+            }
             const sponsorData = {
-                reason,
+                organizationName: companyName,
+                responsible: responsibleName,
                 contact: email,
-                logoId: logoId ?? undefined, // si no está cargado, se omite
+                reason,
+                wantsLogo,
+                wantsBanner,
+                logoId,
+                bannerId,
             };
 
             await createSponsor(authToken, sponsorData);
             alert("Solicitud enviada con éxito");
-            // Limpiar todos los campos
+
+            // Limpiar formulario
             setCompanyName('');
             setResponsibleName('');
             setEmail('');
@@ -48,35 +59,12 @@ export default function SponsorFormPage() {
             setLogoFile(null);
             setBannerFile(null);
             setReason('');
-            setLogoPreview(null);
-            setBannerPreview(null);
-            setLogoId(null);
-
 
         } catch (error) {
             console.error("Error al enviar solicitud:", error);
             alert("Error al enviar la solicitud");
         }
-
     };
-    const handleSubmitMedia = async () => {
-        if (!authToken) {
-            alert("el Token de autenticacion es requerido");
-            return;
-        }
-
-        if (wantsLogo && logoFile) {
-            try {
-                const uploadedId = await postMedia(logoFile, authToken);
-                setLogoId(uploadedId);
-            } catch (error) {
-                console.error("Error al subir el logo:", error);
-                alert("Error al subir el logo");
-            }
-        }
-    };
-
-
 
     return (
         <div className="max-w-xl mx-auto p-6 mt-8 bg-white rounded-2xl font-roboto border-0 shadow-none">
@@ -86,46 +74,43 @@ export default function SponsorFormPage() {
                     Estás a un paso de convertirte en un <br />
                     auspiciante y ayudar a nuestra causa
                 </p>
-                <br />
+
                 <label>Nombre de la empresa</label>
                 <input
                     type="text"
                     value={companyName}
                     onChange={(e) => setCompanyName(e.target.value)}
-                    className="w-full p-2 rounded-md border-2 border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-600"
+                    className="w-full p-2 rounded-md border-2 border-blue-500"
                     maxLength={30}
                     required
                 />
-                <br />
-                <br />
+
                 <label>Nombre del responsable</label>
                 <input
                     type="text"
                     value={responsibleName}
                     onChange={(e) => setResponsibleName(e.target.value)}
-                    className="w-full p-2 rounded-md border-2 border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-600"
+                    className="w-full p-2 rounded-md border-2 border-blue-500"
                     maxLength={30}
                     required
                 />
-                <br />
-                <br />
+
                 <label>Correo</label>
                 <input
                     type="email"
                     value={email}
                     onChange={(e) => setEmail(e.target.value)}
-                    className="w-full p-2 rounded-md border-2 border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-600"
+                    className="w-full p-2 rounded-md border-2 border-blue-500"
                     maxLength={30}
                     required
                 />
-                <br />
-                <br />
-                <label>Razón por la cual quiere ser Auspiciante</label>
+
+                <label>Razón por la cual quiere ser Auspiciante:</label>
                 <input
                     type="text"
                     value={reason}
                     onChange={(e) => setReason(e.target.value)}
-                    className="w-full p-2 rounded-md border-2 border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-600"
+                    className="w-full p-2 rounded-md border-2 border-blue-500"
                     maxLength={200}
                     required
                 />
@@ -140,16 +125,16 @@ export default function SponsorFormPage() {
                 </label>
 
                 {wantsLogo && (
-                    <div className="w-64 h-[16rem] text-blue text-2xl rounded-xl overflow-hidden  border-2 border-[rgb(158,189,255)] hover:shadow-md hover:shadow-[rgb(185,207,255)]  flex flex-col relative items-center justify-center mx-auto">
+                    <div className="w-64 h-64 text-blue text-2xl rounded-xl border-2 border-blue-300 flex flex-col items-center justify-center mx-auto">
                         <label className="cursor-pointer block">
                             + Añadir logo
                             <input
                                 type="file"
                                 accept="image/*"
+                                onChange={(e) => setLogoFile(e.target.files?.[0] ?? null)}
                                 className="hidden"
                             />
                         </label>
-                    
                     </div>
                 )}
 
@@ -163,29 +148,21 @@ export default function SponsorFormPage() {
                 </label>
 
                 {wantsBanner && (
-                    <div className="h-[16rem] text-blue text-2xl rounded-xl overflow-hidden  border-2 border-[rgb(158,189,255)] hover:shadow-md hover:shadow-[rgb(185,207,255)]  flex flex-col relative items-center justify-center">
+                    <div className="h-64 text-blue text-2xl rounded-xl border-2 border-blue-300 flex flex-col items-center justify-center">
                         <label className="cursor-pointer block">
                             + Añadir banner
                             <input
                                 type="file"
                                 accept="image/*"
-                                onChange={(e) => {
-                                    const file = e.target.files?.[0] ?? null;
-                                    setBannerFile(file);
-                                    if (file) {
-                                        setBannerPreview(URL.createObjectURL(file));
-                                    }
-                                }}
+                                onChange={(e) => setBannerFile(e.target.files?.[0] ?? null)}
                                 className="hidden"
                             />
-
                         </label>
                         <p className="text-sm text-gray-500 mt-2">Tamaño sugerido: 900x300 píxeles</p>
                     </div>
                 )}
 
                 <div className="flex flex-col justify-center gap-4 mt-6">
-                    {/* Botón primario */}
                     <button
                         type="submit"
                         className="bg-gradient-to-r from-purple-500 to-purple-600 text-white text-lg font-semibold py-3 px-8 rounded-xl shadow-sm hover:from-purple-600 hover:to-purple-700 transition-colors w-fit mx-auto"
@@ -200,7 +177,6 @@ export default function SponsorFormPage() {
                         Ir a Inicio
                     </Link>
                 </div>
-
             </form>
         </div>
     );
